+++ src/translations/types.ts
@@ -20,6 +20,9 @@ export type PluginLanguage = Language<{
     addressPostalCode: string
     addressState: string
     addressTitle: string
+    addressTitleSir?: string
+    addressTitleMadam?: string
+    addressTitleOther?: string
     amount: string
     availableVariants: string
     billing: string